requirements.txt
langchain>=0.3.20
langchain-community>=0.3.19
langchain-groq>=0.2.0
langchain-huggingface>=0.1.2
langchain-chroma>=0.2.2
langchain-text-splitters>=0.3.6
chromadb>=0.6.3
streamlit>=1.42.2
sentence-transformers>=3.4.1
pypdf>=5.3.1
pdfplumber>=0.11.5
python-dotenv>=1.0.1
